@@ -13,13 +13,12 @@
     <RepositoryType>git</RepositoryType>
     <RepositoryUrl>https://github.com/martincostello/SignInWithAppleSample.git</RepositoryUrl>
     <RootNamespace>MartinCostello.SignInWithApple</RootNamespace>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <UserSecretsId>MartinCostello.SignInWithApple</UserSecretsId>
     <VersionPrefix>1.0.0</VersionPrefix>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="AspNet.Security.OAuth.Apple" Version="2.1.1-preview-0886" />
+    <PackageReference Include="AspNet.Security.OAuth.Apple" Version="3.0.0-rc1.19471.47" />
   </ItemGroup>
   <Target Name="AddGeneratedContentItems" BeforeTargets="AssignTargetPaths">
     <ItemGroup>
